Rename base CLI module
